refactor: localQuantumTask and QuantumTaskBatch
